+++ NEWS.md
@@ -1,4 +1,4 @@
-# autograph 0.3.2
+# autograph 0.4.0
 
 ## Package